[PATCH] [10/13] quota-10-inttype
Remove use of 'short' in parameters of functions. 'int' is used instead.

+++ fs/inode.c
@@ -878,9 +878,9 @@ void update_atime (struct inode *inode)
 
 /* Functions back in dquot.c */
 void put_dquot_list(struct list_head *);
-int remove_inode_dquot_ref(struct inode *, short, struct list_head *);
+int remove_inode_dquot_ref(struct inode *, int, struct list_head *);
 
-void remove_dquot_ref(struct super_block *sb, short type)
+void remove_dquot_ref(struct super_block *sb, int type)
 {
 	struct inode *inode;
 	struct list_head *act_head;
